Refactor error handling in platform implementations (Fixes #2).
Using "impl From<T> for DbError" and heavy use of the try! macro, the
platform implementations are now "bubbling up" the original driver
errors instead of panicing and at the same time givin the final user
an informative message about what's going on.

In my case I now obtain this:
DbError { description: "Could not connect to address: localhost", cause: None }

instead of a panic with a generic "Error in the query".